+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Add `MakeTxDeposit` to support THORChain swap transactions
 - changed: Move `makeTx` params types to common
+- changed: (RUNE) Replace shapeshift api with native rpc tx query
 - fixed: Possible error in ETH transactions when calculated gasPrice is below baseFee
 
 ## 2.11.2 (2023-11-13)